% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/92cons.r
\name{taiwanConsensus}
\alias{taiwanConsensus}
\title{Taiwan consensus}
\usage{
taiwanConsensus(numCountries, numSystems)
}
\arguments{
\item{numCountries}{Number of countries}

\item{numSystems}{Number of systems}
}
\value{
character
}
\description{
Taiwan consensus
}
